requirements.txt
# Microsoft Agent Framework (https://github.com/microsoft/agent-framework)
agent-framework --pre
agent-framework-orchestrations --pre
agent-framework-foundry-local --pre

# Foundry Local SDK (https://foundrylocal.ai / https://github.com/microsoft/foundry-local)
foundry-local-sdk

# UI server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0

# Testing
pytest>=8.0.0
